rag_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.2
qdrant-client==1.7.0
asyncpg==0.29.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.12
python-dotenv==1.0.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2